+++ CHANGES.rst
@@ -4,6 +4,8 @@ dirsync - Changes
 2.3 (dev)
 ---------
 
+- Added comparing content of files
+
 
 2.2 (28-01-2015)
 ----------------